Refactored project names, namespaces, folders, etc
