+++ .template.config/template.json
@@ -2,13 +2,13 @@
     "$schema": "http://json.schemastore.org/template",
     "author": "Hubert Graham",
     "classifications": [ ".NET 6 API",  "MVC", "Template"],
-    "identity": "Kode4Hue.CleanArchitecture.API",
-    "shortName": "Kode4Hue.CleanArchitecture.API",
-    "name": "Kode4Hue.CleanArchitecture.API",
+    "identity": "Kode4Hue.API",
+    "shortName": "Kode4Hue.API",
+    "name": "Kode4Hue.API",
     "tags": {
 	    "language": "C#",
 	    "type": "project"
 	  },
-    "sourceName": "CleanArchitecture.API",
+    "sourceName": "API",
     "preferNameDirectory": true
 }
