+++ API.Presentation/Dockerfile
@@ -1,0 +1,22 @@
+#See https://aka.ms/containerfastmode to understand how Visual Studio uses this Dockerfile to build your images for faster debugging.
+
+FROM mcr.microsoft.com/dotnet/aspnet:6.0 AS base
+WORKDIR /app
+EXPOSE 80
+EXPOSE 443
+
+FROM mcr.microsoft.com/dotnet/sdk:6.0 AS build
+WORKDIR /src
+COPY ["API.Presentation/API.Presentation.csproj", "API.Presentation/"]
+RUN dotnet restore "API.Presentation/API.Presentation.csproj"
+COPY . .
+WORKDIR "/src/API.Presentation"
+RUN dotnet build "API.Presentation.csproj" -c Release -o /app/build
+
+FROM build AS publish
+RUN dotnet publish "API.Presentation.csproj" -c Release -o /app/publish
+
+FROM base AS final
+WORKDIR /app
+COPY --from=publish /app/publish .
+ENTRYPOINT ["dotnet", "API.Presentation.dll"]
