+++ Application/Common/Extensions/ServicesExtensions.cs
@@ -1,0 +1,23 @@
+﻿using MediatR;
+using Microsoft.Extensions.DependencyInjection;
+using System.Reflection;
+using FluentValidation;
+using CleanArchitecture.SharedLibrary.Common.Behaviours;
+
+namespace Application.Common.Extensions
+{
+    public static class ServicesExtensions
+    {
+        public static IServiceCollection AddApplication(this IServiceCollection services)
+        {
+            var executingAssembly = Assembly.GetExecutingAssembly();
+            services.AddValidatorsFromAssembly(executingAssembly);
+            services.AddMediatR(executingAssembly);
+            services.AddTransient(typeof(IPipelineBehavior<,>), typeof(UnhandledExceptionBehaviour<,>));
+            services.AddTransient(typeof(IPipelineBehavior<,>), typeof(ValidationBehaviour<,>));
+            services.AddTransient(typeof(IPipelineBehavior<,>), typeof(PerformanceBehaviour<,>));
+
+            return services;
+        }
+    }
+}
