+++ Application/PetrolPrices/DTOs/OutletDto.cs
@@ -1,0 +1,8 @@
+﻿namespace Application.PetrolPrices.DTOs
+{
+    public class OutletDto
+    {
+        public string Id { get; set; }
+        public string Name { get; set; }
+    }
+}
