+++ Application/PetrolPrices/Queries/GetPetrolPrices/GetPetrolPricesQuery.cs
@@ -1,0 +1,12 @@
+﻿using Application.PetrolPrices.DTOs;
+using CleanArchitecture.SharedLibrary.Common.DTOs;
+using MediatR;
+
+namespace Application.PetrolPrices.Queries.GetPetrolPrices
+{
+    public class GetPetrolPricesQuery : IRequest<PaginatedList<PetrolPriceDto>>
+    {
+        public int Limit { get; set; }
+        public int PageNumber { get; set; }
+    }
+}
