--- CleanArchitecture.API.Presentation/Common/Controllers/ApiBaseController.cs
@@ -1,17 +1,0 @@
-﻿using MediatR;
-using Microsoft.AspNetCore.Mvc;
-
-namespace CleanArchitecture.API.Presentation.Common.Controllers
-{
-    [ApiController]
-    [Route("api/[controller]/[action]")]
-    public class ApiBaseController : ControllerBase
-    {
-        protected ISender Mediator;
-
-        public ApiBaseController(ISender mediator)
-        {
-            Mediator = mediator;
-        }
-    }
-}
