--- CleanArchitecture.Application/PetrolPrices/DTOs/OutletDto.cs
@@ -1,8 +1,0 @@
-﻿namespace CleanArchitecture.Application.PetrolPrices.DTOs
-{
-    public class OutletDto
-    {
-        public string Id { get; set; }
-        public string Name { get; set; }
-    }
-}
