--- CleanArchitecture.Infrastructure/Extensions/ServicesExtensions.cs
@@ -1,28 +1,0 @@
-﻿using CleanArchitecture.Application.Services.PetrolPrices;
-using CleanArchitecture.Infrastructure.PetrolPrices.Services;
-using JamaicaOpenData.SharedLibrary.Common.Extensions;
-using Microsoft.Extensions.DependencyInjection;
-using System.Reflection;
-
-namespace CleanArchitecture.Infrastructure.Extensions
-{
-    public static class ServicesExtensions
-    {
-        public static IServiceCollection AddInfrastructure(this IServiceCollection services)
-        {
-            services.AddMappingProfiles();
-            services.RegisterJamaicaOpenDataServices();
-            services.AddTransient<IPetrolPriceService, PetrolPriceService>();
-            return services;
-        }
-
-        private static IServiceCollection AddMappingProfiles(this IServiceCollection services)
-        {
-            services.AddAutoMapper(config =>
-            {
-                config.AddMaps(Assembly.GetExecutingAssembly());
-            });
-            return services;
-        }
-    }
-}
