+++ README.md
@@ -1,4 +1,4 @@
-# Kode4Hue.CleanArchitecture.API
+# Kode4Hue.API
 
 A .NET 6 MVC API template based on the principles of clean architecture. Feel free to grab it and use it on your next .NET API project.
 
@@ -8,19 +8,19 @@ A .NET 6 MVC API template based on the principles of clean architecture. Feel fr
 
 ### Getting Started:
 * Open a Terminal Window and set the current directory to the root directory of this template.
-* Run `dotnet new --install Kode4Hue.CleanArchitecture.API` to install the API template
+* Run `dotnet new --install Kode4Hue.API` to install the API template
 * Confirm the template is installed by running the `dotnet new --list` command to view all installed templates
-* Once the template is installed, you then can use the template either from Visual Studio or using the dotnet cli by running: `dotnet new Kode4Hue.CleanArchitecture.API -o YourProjectName`
+* Once the template is installed, you then can use the template either from Visual Studio or using the dotnet cli by running: `dotnet new Kode4Hue.API -o YourProjectName`
 
 ### Clean Architecture Layers:
-* Presentation - CleanArchitecture.API.Presentation
-* Infrastructure - CleanArchitecture.Infrastructure
-* Application - CleanArchitecture.Application
-* Domain - CleanArchitecture.Domain
+* Presentation - API.Presentation
+* Infrastructure - Infrastructure
+* Application - Application
+* Domain - Domain
 
 ### Core Technology Stack:
 * [ASP.NET Core 6](https://docs.microsoft.com/en-us/aspnet/core/introduction-to-aspnet-core?view=aspnetcore-6.0)
-* [Kode4Hue.CleanArchitecture.SharedLibrary](https://github.com/Kode4Hue/CleanArchitecture.SharedLibrary)
+* [Kode4Hue.CleanArchitecture.SharedLibrary](https://github.com/Kode4Hue/SharedLibrary)
 * [Kode4Hue.JamaicaOpenData.SharedLibrary](https://github.com/Kode4Hue/JamaicaOpenData.SharedLibrary)
 * [MediatR](https://github.com/jbogard/MediatR)
 * [FluentValidation](https://fluentvalidation.net/)
